% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.eNewick.R
\name{convert.eNewick}
\alias{convert.eNewick}
\title{Converts Extended Newick Format to Traditional Newick Format}
\usage{
convert.eNewick(eNewick)
}
\arguments{
\item{eNewick}{phylogenetic tree in extended Newick format.}
}
\value{
phylogenetic tree in traditional Newick format.
}
\description{
Takes a phylogenetic tree in extended Newick format and converts it to
traditional Newick format that can be directly manipulated by packages such
as \code{ape} and \code{phytools}.
}
\examples{
### Converts the phylogenetic tree into traditional Newick format.
 tree <- "((A:0.1,(B:0.3,C:0.2):0.2[60]):0.4[100],(E:0.12,F:0.09):0.4[100]);"
 new.tree <- convert.eNewick(tree)
 new.tree
}

