\name{PhySortR-package}
\alias{PhySortR-package}
\docType{package}
\title{
A Fast, Flexible Tool for Sorting Phylogenetic Trees
}
\description{
  \strong{PhySortR} provides a quick and highly flexible function for the screening (within a tree) of Exclusive clades that comprise only the target taxa and/or Non-Exclusive clades that includes a defined portion of non-target taxa. Support is also provided for both traditional and extended Newick formatted phylogenetic trees.

A full list of functions can be displayed by \code{library(help = PhySortR)}.
}
\details{
  \tabular{ll}{
  Package: \tab PhySortR\cr
  Type: \tab Package\cr
  Version: \tab 1.0.7\cr
  Date: \tab 2016-05-16\cr
  License: \tab GPL (>= 3)\cr
  }
}
\author{
  Timothy G. Stephens, Debashish Bhattacharya, Mark A. Ragan, Cheong Xin Chan
  
  Maintainer: Timothy G. Stephens <timothy.stephens1@uqconnect.edu.au>
}
\references{
  Stephens TG, Bhattacharya D, Ragan MA, Chan CX. 2016. PhySortR: a fast, flexible tool for sorting phylogenetic trees in R. PeerJ, 4:e2038, DOI:10.7717/peerj.2038
}
\keyword{ package }
\seealso{
  \code{
    \link{phytools}
    \link{ape}
  }
}